+++ lib/src/theme/theme.dart
@@ -1,3 +1,5 @@
 export 'filter_list_theme.dart';
 export 'choice_chip_theme.dart';
 export 'header_theme.dart';
+export 'contol_button_theme.dart';
+export 'control_button_bar_theme.dart';