+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - Add `FilterListTheme` for filter list widget theme
 - Add `ChoiceChipTheme` for choice chip theme.
 - Add `HeaderTheme` for Header widget theme
+- Add `ControlButtonBarTheme` for control button bar theme
+- Add `ControlButtonTheme` for control button theme
 
 ## [1.0.0] 21 May 2021